rename mint service to usdcDemo
